pyproject.toml
[project]
name = "llm-mail-description"
version = "0.1.0"
description = "Email analysis service using local LLM with constrained JSON output"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.2",
    "jinja2>=3.1.6",
    "httpx>=0.28.0",
    "tritonclient[grpc]>=2.64.0",
    "outlines>=1.2.0",
    "structlog>=25.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "mypy>=1.15.0",
    "types-PyYAML>=6.0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
